Added capabilities to load level from file
The format spec of the loader file will be added to the readme after it has been completed. As of currently, it is possible to add a player, world boundaries, the water body, the player spawn islands, collectables, and a park keeper (to be dumb ai) which currently does nothing.